Expose `WithLifetime` trait under a new `advanced` module, and move arities & type_eq to it
